サーバーのモジュール化
