+++ tsconfig.app.json
@@ -27,8 +27,7 @@
     "src/**/*.tsx",
     "src/**/*.vue",
     "prisma/seed.ts",
-    "src/api/server.ts",
-    "src/api/create_user_api.ts",
+    "src/api/user/create.ts",
     "src/api/achievement/getItems.ts"
   ]
 }
